Refactored series into a single file and moved stats.js to lib/
