+++ lib/stats.js
@@ -7,82 +7,13 @@ var Promise       = require('promise');
 var request       = require('superagent-promise');
 var url           = require('url');
 var urljoin       = require('url-join');
+var series        = require('./series');
 
-/** Types supported for reporting */
-var types = {
-  String: function(v) { return /string/.test(typeof(v)); },
-  Number: function(v) { return /number/.test(typeof(v)); },
-  Any:    function(v) { return /string|number/.test(typeof(v)); }
-};
+// Export types allowed for series
+exports.types = series.types;
 
-// Export types
-exports.types = types;
-
-/**
- * Create a times series that you can report to.
- *
- * options:
- * {
- *   // Name of the series in influxdb
- *   name:               'ResponseTimes',
- *
- *   // Required columns
- *   columns: {
- *     method:           stats.types.String,
- *     duration:         stats.types.Number,
- *     custom:           stats.types.Any
- *   },
- *
- *   // Type of additional columns if they are allowed (defaults `null`)
- *   additionalColumns:  stats.types.String
- * }
- */
-var Series = function(options) {
-  // Validate options
-  assert(options,         "options are required");
-  assert(options.name,    "Series must be named");
-  options = _.defaults({}, options, {
-    columns:              {},
-    additionalColumns:    null
-  });
-
-  // Store options
-  this._options = options;
-};
-
-// Export series
-exports.Series = Series;
-
-/**
- * Create a reporter function that will validate points and submit them to the
- * drain.
- */
-Series.prototype.reporter = function(drain) {
-  var options = this._options;
-
-  // Return a reporter
-  return function(point) {
-    // Validate that
-    _.forIn(point, function(value, key) {
-      var validate = options.columns[key] || options.additionalColumns;
-      if (!validate) {
-        debug("additionalColumn %s not allowed in series: %s",
-              key, options.name);
-        throw new Error("additionalColumn " + key + " is not allowed!");
-      }
-      if (!validate(value)) {
-        debug("Failed validate %s for key %s in series: %s",
-              value, key, options.name);
-        throw new Error("Failed validate key " + key + " in series: " +
-                        options.name);
-      }
-    });
-
-    // If validated we'll add the point to the drain
-    drain.addPoint(options.name, point);
-  };
-};
-
+// Export series constructor
+exports.Series = series.Series;
 
 /**
  * Create an Influx Database Connection
@@ -368,20 +299,6 @@ var createResponseTimer = function(reporter, additionalValues) {
 // Export createResponseTimer
 exports.createResponseTimer = createResponseTimer;
 
-
-/** Handler reports format for createHandlerTimer */
-var HandlerReports = new Series({
-  name:       'HandlerReports',
-  columns: {
-    component:      types.String,
-    duration:       types.Number,
-    exchange:       types.String,
-    redelivered:    types.String,   // true || false
-    error:          types.String    // true || false
-  }
-});
-
-
 /**
  * Create a handler timer for AMQP messages received through
  * taskcluster-client. Please note, that this relies on that messages format.
@@ -399,7 +316,7 @@ var createHandlerTimer = function(handler, options) {
   assert(options.component,           "options.component is required");
 
   // Create a reporter
-  var reporter = HandlerReports.reporter(options.drain);
+  var reporter = series.HandlerReports.reporter(options.drain);
 
   // Wrap handler and let that be it
   return function(message) {
@@ -448,17 +365,6 @@ var createHandlerTimer = function(handler, options) {
 exports.createHandlerTimer = createHandlerTimer;
 
 
-/** Usage reports format for monitorProcessUsage */
-var UsageReports = new Series({
-  name:       'UsageReports',
-  columns: {
-    component:      types.String,
-    process:        types.String,
-    cpu:            types.Number,
-    memory:         types.Number
-  }
-});
-
 /** Interval handle for process usage monitoring */
 var _processUsageReportingInterval = null;
 
@@ -497,7 +403,7 @@ var startProcessUsageReporting = function(options) {
   var usage = require('usage');
 
   // Create reporter
-  var reporter = UsageReports.reporter(options.drain);
+  var reporter = series.UsageReports.reporter(options.drain);
 
   // Set interval to report usage at interval
   _processUsageReportingInterval = setInterval(function() {
